requirements.txt.txt
streamlit
pandas
plotly
xhtml2pdf
fpdf2
reportlab